backend/requirements.txt
fastapi
uvicorn
numpy
pandas
requests
python-multipart
pydantic
tensorflow-macos
tensorflow-metal
pyarrow
recommenders